MainCell 코드 분리
