+++ ios/ToDoListApp/ToDoListApp/Presentation/UI/Views/TableViewDragDelegate.swift
@@ -1,0 +1,29 @@
+//
+//  TableviewDragDelegate.swift
+//  ToDoListApp
+//
+//  Created by zombietux on 2021/04/13.
+//
+
+import UIKit
+
+class TableViewDragDelegate: NSObject {
+    
+}
+
+extension TableViewDragDelegate: UITableViewDragDelegate, UITableViewDropDelegate {
+    
+    func tableView(_ tableView: UITableView, itemsForBeginning session: UIDragSession, at indexPath: IndexPath) -> [UIDragItem] {
+        return [UIDragItem(itemProvider: NSItemProvider())]
+    }
+    
+    func tableView(_ tableView: UITableView, dropSessionDidUpdate session: UIDropSession, withDestinationIndexPath destinationIndexPath: IndexPath?) -> UITableViewDropProposal {
+        if session.localDragSession != nil {
+            return UITableViewDropProposal(operation: .move, intent: .insertAtDestinationIndexPath)
+        }
+        return UITableViewDropProposal(operation: .cancel, intent: .unspecified)
+    }
+    
+    func tableView(_ tableView: UITableView, performDropWith coordinator: UITableViewDropCoordinator) {}
+}
+
